requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torch-geometric>=2.3.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.5

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0
pingouin>=0.5.3
lifelines>=0.27.0

# Data Manipulation and Analysis
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0.0

# Cheminformatics
rdkit-pypi>=2022.9.5
mordred>=1.2.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
bokeh>=3.2.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0

# CLI Development
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
tqdm>=4.65.0

# Model Interpretation
shap>=0.42.0
lime>=0.2.0.1

# Data Validation and Processing
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing and Code Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Notebook and Documentation
jupyter>=1.0.0
jupyterlab>=4.0.0
nbconvert>=7.0.0

# Database and Caching
sqlalchemy>=2.0.0
redis>=4.6.0

# Deployment and MLOps
docker>=6.1.0
mlflow>=2.5.0

# Additional Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0